requirements.txt
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
plotly==5.24.1
PyPDF2==3.0.1
python_docx==1.1.2
seaborn==0.13.2
streamlit==1.37.1
wordcloud==1.9.4
